# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation about projekt SasiadWPotrzebie",
        "description": "sasiadwpotrzebie api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "local server"
        },
        {
            "url": "http://localhost:9000",
            "description": "staging server"
        },
        {
            "url": "http://localhost:9000",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Retrieve all cities",
                "description": "Get all cities",
                "operationId": "702928da5dbd036fe945b2c774e18a3f",
                "responses": {
                    "200": {
                        "description": "List of all cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Warsaw"
                                            },
                                            "commune_id": {
                                                "type": "integer",
                                                "example": 101
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-22T10:20:30Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-23T15:30:45Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/cities/{id}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Retrieve a city by its ID",
                "description": "Get city by id",
                "operationId": "7368586a528b201d1537cd877dd5cca7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Warsaw"
                                        },
                                        "commune_id": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-22T10:20:30Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-23T15:30:45Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "City not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/communes/{id}/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Retrieve cities by commune ID",
                "description": "Get cities by commune",
                "operationId": "67c4451ea3e5d0f65fa06b78cfb2947d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commune ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cities in the commune",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Krakow"
                                            },
                                            "commune_id": {
                                                "type": "integer",
                                                "example": 102
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-25T11:15:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-26T17:45:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commune not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}